+++ Changelog.md
@@ -1,5 +1,10 @@
 # Motoko compiler changelog
 
+* motoko (`moc`)
+
+  * improvement: RTS weak reference interaction with the incremental GC: weak
+    reference reads now go through a load barrier (#6296).
+
 ## 1.14.0 (2026-08-11)
 
 * motoko (`moc`)